Download the Time Zone Database release files when building and testing.
Download the tzdata release when building modules. Download the tzcode
and tzdata releases and compile zdump and zic when running tests.

Downloads, extracted files and compiled executables are cached in the
tzdb directory.

Environment variables can be set to prevent the downloads: TZDATA
specifies the path to a directory containing extracted tzdata release.
ZDUMP and ZIC specify the path to the zdump and zic executables to
use to run tests. It is recommended that the zdump and zic versions
match the tzdata release.

Remove the extracted tzdata release files from the tzinfo-data
repository.